Move Sauna and Room objects to data_models/
