+++ src/smart_sauna_map/data_models/room.py
