+++ src/smart_sauna_map/searcher/sauna_ikitai_searcher.py
@@ -8,9 +8,9 @@ from urllib.parse import urlencode, urlparse
 import requests
 from bs4 import BeautifulSoup
 
+from smart_sauna_map.data_models.room import MansRoom, UnisexRoom, WomansRoom
+from smart_sauna_map.data_models.sauna import Sauna
 from smart_sauna_map.geocoding import geocode
-from smart_sauna_map.room import MansRoom, UnisexRoom, WomansRoom
-from smart_sauna_map.sauna import Sauna
 from smart_sauna_map.searcher.abstract_searcher import AbstractSearcher
